Niklas' comments addressed.
Minor: It looks like you could delete the constructor and destructor for DiretXTileRenderer, since they don't do anything anyway.
Abi: deleted them.
In DirecxTileRenderer::Initialize, you assign the return values of CreateFactory and CreateDevice to const references to com_ptr. You should remove the ampersand. Making it a reference doesn't serve any purpose, and in fact I'm surprised it works, since the returned temporary com_ptr that gets bound to the reference should be destroyed at the end of the full expression.
Abi: changed it to "auto"
Minor: In DirectXTileRenderer::DrawTileRange, it would be simpler and clearer to remove the it local variable and used a ranged for loop: for (Tile& tile : tiles) { ... }.
Abi: Looks cleaner for sure. Changed it.
In DirectXTileRenderer::DrawTile, if you're creating a new solid color brush inside every call, the new brush should be assigned to a local com_ptr, not to the m_textBrush member. Alternatively, create the m_textBrush once in Initialize and call SetColor on it for each tile you draw.
Abi Good catch, fixed it.
Are you sure device-lost is handled correctly? It seems like you create a bunch of device-dependent resources inside Initialize. Does Initialize get called again if the device is lost?
Abi Not entirely sure, I think I need to do this right again. May be in the next PR.
Minor: The alignment of your member variables in TileDrawingManager.h is odd. Is this a tabs vs. spaces issue, or something else?
Abi I used only tabs. Maybe the environment is different.
There's a TODO comment in WinComp::ConfigureInteraction. Did you intend to resolve that before publishing the sample?
Abi Most likely not, since we are running out of time, but we ll fix this after this PR.